refactor(ui): extract embed views and utilities from blocks-content.tsx
Move reusable exports out of the ~2800-line blocks-content.tsx into
focused files to reduce coupling and prepare for future migration:

- embed-views.tsx: BlockEmbedCard, BlockEmbedContent, BlockEmbedComments,
  BlockEmbedContentComment, ErrorBlock, DeletedEmbedBanner
- blocks-content-utils.tsx: getBlockNodeById, getBlockNode, blockStyles,
  DocumentCardGrid, collectImageBlocks, resolveGalleryNavigation,
  resolveSwipeDirection, ImageGalleryItem type
- Move BlockRangeSelectOptions type to @shm/shared/document-content-props

blocks-content.tsx retains re-exports for backward compatibility.
Consumer imports updated to use the new paths directly.
