+++ frontend/packages/ui/src/comments.tsx
@@ -42,7 +42,9 @@ import {Link, MessageSquare, Pencil, Trash2, X} from 'lucide-react'
 import {memo, ReactNode, useEffect, useLayoutEffect, useMemo, useRef, useState} from 'react'
 import {toast} from 'sonner'
 import {SelectionContent} from './accessories'
-import {BlockRangeSelectOptions, BlocksContent, BlocksContentProvider, getBlockNodeById} from './blocks-content'
+import type {BlockRangeSelectOptions} from '@shm/shared/document-content-props'
+import {BlocksContent, BlocksContentProvider} from './blocks-content'
+import {getBlockNodeById} from './blocks-content-utils'
 import {Button} from './button'
 import {Popover, PopoverContent, PopoverTrigger} from './components/popover'
 import {copyTextToClipboard, copyUrlToClipboardWithFeedback} from './copy-to-clipboard'
