+++ frontend/packages/ui/src/resource-page-common.tsx
@@ -1,4 +1,5 @@
 import {BlockRange, HMDocument, HMExistingDraft, UnpackedHypermediaId} from '@seed-hypermedia/client/hm-types'
+import type {DocumentContentProps} from '@shm/shared/document-content-props'
 import {
   createInspectNavRoute,
   DocumentPanelRoute,
@@ -43,7 +44,7 @@ import {useNavigate, useNavRoute} from '@shm/shared/utils/navigation'
 import {Folder, Search} from 'lucide-react'
 import {CSSProperties, lazy, ReactNode, Suspense, useCallback, useEffect, useMemo, useRef, useState} from 'react'
 import {AccountPage} from './account-page'
-import {BlockRangeSelectOptions, BlocksContent, BlocksContentProvider} from './blocks-content'
+import type {BlockRangeSelectOptions} from '@shm/shared/document-content-props'
 import {CollaboratorsPage} from './collaborators-page'
 import {ScrollArea} from './components/scroll-area'
 import {copyUrlToClipboardWithFeedback} from './copy-to-clipboard'
@@ -206,6 +207,8 @@ export interface ResourcePageProps {
   inspect?: (inspectionEvent: any) => void
   /** Inspect event store for the debug drawer to subscribe to. */
   inspectStore?: import('@shm/shared/models/document-machine-inspect').InspectEventStore
+  /** Component to render document content using the editor. Optional during SSR. */
+  DocumentContentComponent?: React.ComponentType<DocumentContentProps>
 }
 
 /** Get panel title for display */
@@ -242,6 +245,7 @@ export function ResourcePage({
   machine,
   inspect,
   inspectStore,
+  DocumentContentComponent,
 }: ResourcePageProps) {
   const route = useNavRoute()
   const isSiteProfile = route.key === 'site-profile'
@@ -411,6 +415,7 @@ export function ResourcePage({
             CommentEditor={CommentEditor}
             siteUrl={siteHomeDocument?.metadata?.siteUrl}
             pageFooter={pageFooter}
+            DocumentContentComponent={DocumentContentComponent}
           />
         </PageWrapper>
       </DocumentMachineProvider>
@@ -461,6 +466,7 @@ export function ResourcePage({
           pageFooter={pageFooter}
           inlineCards={inlineCards}
           inlineInsert={inlineInsert}
+          DocumentContentComponent={DocumentContentComponent}
         />
       </PageWrapper>
       {inspect && (
@@ -600,6 +606,7 @@ function DocumentBody({
   inlineCards,
   openComment,
   inlineInsert,
+  DocumentContentComponent,
 }: {
   docId: UnpackedHypermediaId
   document: HMDocument
@@ -618,6 +625,8 @@ function DocumentBody({
   openComment?: string
   /** Optional inline element injected after content blocks */
   inlineInsert?: ReactNode
+  /** Component to render document content using the editor */
+  DocumentContentComponent?: React.ComponentType<DocumentContentProps>
 }) {
   // Sync document into state machine
   useDocumentSync(document)
@@ -1142,6 +1151,7 @@ function DocumentBody({
           siteUrl={siteUrl}
           inlineCards={inlineCards}
           inlineInsert={inlineInsert}
+          DocumentContentComponent={DocumentContentComponent}
         />
       </div>
       {pageFooter ? <div className="mt-auto">{pageFooter}</div> : null}
@@ -1339,6 +1349,7 @@ function MainContent({
   siteUrl,
   inlineCards,
   inlineInsert,
+  DocumentContentComponent,
 }: {
   docId: UnpackedHypermediaId
   resourceId: UnpackedHypermediaId
@@ -1375,6 +1386,7 @@ function MainContent({
   siteUrl?: string
   inlineCards?: ReactNode
   inlineInsert?: ReactNode
+  DocumentContentComponent?: React.ComponentType<DocumentContentProps>
 }) {
   switch (activeView) {
     case 'directory':
@@ -1450,6 +1462,7 @@ function MainContent({
           directory={directory}
           inlineCards={inlineCards}
           inlineInsert={inlineInsert}
+          DocumentContentComponent={DocumentContentComponent}
         />
       )
   }
@@ -1471,6 +1484,7 @@ function ContentViewWithOutline({
   directory,
   inlineCards,
   inlineInsert,
+  DocumentContentComponent,
 }: {
   docId: UnpackedHypermediaId
   resourceId: UnpackedHypermediaId
@@ -1491,6 +1505,7 @@ function ContentViewWithOutline({
   directory?: import('@seed-hypermedia/client/hm-types').HMDocumentInfo[]
   inlineCards?: ReactNode
   inlineInsert?: ReactNode
+  DocumentContentComponent?: React.ComponentType<DocumentContentProps>
 }) {
   const outline = useNodesOutline(document, docId)
 
@@ -1519,16 +1534,18 @@ function ContentViewWithOutline({
       )}
 
       <div {...mainContentProps}>
-        <BlocksContentProvider
-          resourceId={resourceId}
-          blockCitations={blockCitations}
-          onBlockCitationClick={onBlockCitationClick}
-          onBlockCommentClick={onBlockCommentClick}
-          onBlockSelect={onBlockSelect}
-          inlineInsert={inlineInsert}
-        >
-          <BlocksContent blocks={document.content} />
-        </BlocksContentProvider>
+        {DocumentContentComponent ? (
+          <DocumentContentComponent
+            blocks={document.content}
+            resourceId={resourceId}
+            focusBlockId={docId.blockRef ?? undefined}
+            blockCitations={blockCitations}
+            onBlockCitationClick={onBlockCitationClick}
+            onBlockCommentClick={onBlockCommentClick}
+            onBlockSelect={onBlockSelect}
+          />
+        ) : null}
+        {inlineInsert}
         {inlineCards}
         <UnreferencedDocuments docId={docId} content={document.content} directory={directory} />
       </div>
